Move some scripts around for now
