+++ tests/conftest.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 import boa.__main__ as dunder_main
-import boa.test_scripts.run_branin as run_branin
+import boa.scripts.run_branin as run_branin
 from boa import cd_and_cd_back, load_yaml, split_shell_command
 from boa.definitions import ROOT
